Moves all models into a single file - makes life easier in shells, with Alembic etc
